+++ ICD.Connect.Misc.CrestronPro/Ports/IrPort/IrPortAdapter.cs
@@ -4,6 +4,7 @@ using ICD.Common.Utils;
 using ICD.Common.Utils.Timers;
 using ICD.Connect.API.Commands;
 using ICD.Connect.API.Nodes;
+using ICD.Connect.Misc.CrestronPro.Utils;
 using ICD.Connect.Protocol.Ports.IrPort;
 using ICD.Connect.Settings.Core;
 #if SIMPLSHARP
@@ -111,10 +112,8 @@ namespace ICD.Connect.Misc.CrestronPro.Ports.IrPort
 		/// <param name="port"></param>
 		private void Unregister(IROutputPort port)
 		{
-			if (port == null || !port.Registered)
-				return;
-
-			port.UnRegister();
+			if (port != null)
+				PortDeviceUtils.Unregister(port);
 		}
 
 		/// <summary>
@@ -123,16 +122,15 @@ namespace ICD.Connect.Misc.CrestronPro.Ports.IrPort
 		/// <param name="port"></param>
 		private void Register(IROutputPort port)
 		{
-			if (port == null)
-				return;
-
-			GenericDevice parent = port.Parent as GenericDevice;
-			if (parent == null)
-				return;
-
-			eDeviceRegistrationUnRegistrationResponse parentResult = parent.ReRegister();
-			if (parentResult != eDeviceRegistrationUnRegistrationResponse.Success)
-				Log(eSeverity.Error, "Unable to register parent {0} - {1}", parent.GetType().Name, parentResult);
+			try
+			{
+				if (port != null)
+					PortDeviceUtils.Register(port);
+			}
+			catch (InvalidOperationException e)
+			{
+				Log(eSeverity.Error, "Error registering port - {0}", e.Message);
+			}
 		}
 #endif